Addressing PR review comments.
