--- cpg_infra/driver/sm_accessor_membership.py
@@ -1,14 +1,0 @@
-# flake8: noqa: ANN401
-"""
-SampleMetadataAccessorMembership named tuple.
-"""
-
-from __future__ import annotations
-
-from typing import Any, Iterable, NamedTuple
-
-
-class SampleMetadataAccessorMembership(NamedTuple):
-    name: str
-    member: Any
-    permissions: Iterable[str]
